../../../openFrameworks/addons/ofxCv/libs/ofxCv/include/ofxCv/Tracker.h:324:35: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]     for(int i = 0; i < labels.size(); i++) {
../../../openFrameworks/addons/ofxCv/libs/ofxCv/include/ofxCv/Tracker.h:324:35: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
    for(int i = 0; i < labels.size(); i++) {

../../../openFrameworks/addons/ofxCv/libs/ofxCv/src/Utilities.cpp:38:36: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   for(int i = 0; i < polyline.size(); i++) {

../../../openFrameworks/addons/ofxCv/libs/ofxCv/src/Utilities.cpp:47:34: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   for(int i = 0; i < points.size(); i++) {

../../../openFrameworks/addons/ofxCv/libs/ofxCv/src/Utilities.cpp:56:34: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   for(int i = 0; i < points.size(); i++) {

../../../openFrameworks/addons/ofxCv/libs/ofxCv/src/ObjectFinder.cpp:84:27: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   for(int i = 0; i < size(); i++) {

../../../openFrameworks/addons/ofxCv/libs/ofxCv/src/Helpers.cpp:77:33: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   for(int i = 0; i < lines.size(); i++) {

../../../openFrameworks/addons/ofxCv/libs/ofxCv/src/Helpers.cpp:101:23: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
    for(int i = 0; i < maxIterations; i++) {

../../../openFrameworks/addons/ofxCv/libs/ofxCv/src/Helpers.cpp:131:33: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   for(int i = 0; i < lines.size(); i++) {

../../../openFrameworks/addons/ofxCv/libs/ofxCv/src/ObjectFinder.cpp:56:35: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   for(int i = 0; i < objects.size(); i++) {

../../../openFrameworks/addons/ofxCv/libs/ofxCv/src/ObjectFinder.cpp:84:27: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
   for(int i = 0; i < size(); i++) {

+++ libs/ofxCv/src/ObjectFinder.cpp
@@ -53,7 +53,7 @@ namespace ofxCv {
                                     (findBiggestObject ? CASCADE_FIND_BIGGEST_OBJECT | CASCADE_DO_ROUGH_SEARCH : 0),
                                     minSize,
                                     maxSize);
-		for(int i = 0; i < objects.size(); i++) {
+		for(size_t i = 0; i < objects.size(); i++) {
             cv::Rect& rect = objects[i];
 			rect.width /= rescale, rect.height /= rescale;
 			rect.x /= rescale, rect.y /= rescale;
@@ -81,7 +81,7 @@ namespace ofxCv {
 	void ObjectFinder::draw() const {
 		ofPushStyle();
 		ofNoFill();
-		for(int i = 0; i < size(); i++) {
+		for(size_t i = 0; i < size(); i++) {
 			ofRectangle object = getObject(i);
 			ofDrawRectangle(object);
 			ofDrawBitmapStringHighlight(ofToString(getLabel(i)), object.x, object.y);
